backend/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
fastapi-jwt-auth==0.5.0
frozendict==2.4.6
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
numpy==2.1.3
orjson==3.10.11
pandas==2.2.3
peewee==3.17.7
platformdirs==4.3.6
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.1
yfinance==0.2.48
